requirements.txt
pip==19.3.1
Flask==1.1.1
Flask-SQLAlchemy===2.1 
Flask-Migrate==1.8.0
psycopg2-binary